+++ cmp/loadingindicator/LoadingIndicator.ts
@@ -7,8 +7,15 @@
 import {hbox} from '@xh/hoist/cmp/layout';
 import {div} from '@xh/hoist/cmp/layout/Tags';
 import {spinner as spinnerCmp} from '@xh/hoist/cmp/spinner';
-import {hoistCmp, HoistModel, HoistProps, Some, TaskObserver, useLocalModel} from '@xh/hoist/core';
-import '@xh/hoist/mobile/register';
+import {
+    Corner,
+    hoistCmp,
+    HoistModel,
+    HoistProps,
+    Some,
+    TaskObserver,
+    useLocalModel
+} from '@xh/hoist/core';
 import {withDefault} from '@xh/hoist/utils/js';
 import classNames from 'classnames';
 import {truncate} from 'lodash';
@@ -18,7 +25,7 @@ export interface LoadingIndicatorProps extends HoistProps {
     /** TaskObserver(s) that should be monitored to determine if the Indicator should be displayed. */
     bind?: Some<TaskObserver>;
     /** Position of the indicator relative to its containing component. */
-    corner?: 'tl' | 'tr' | 'bl' | 'br';
+    corner?: Corner;
     /** True to display the indicator. */
     isDisplayed?: boolean;
     /** Max characters allowed in message, after which it will be elided. Default 30. */
@@ -32,7 +39,7 @@ export interface LoadingIndicatorProps extends HoistProps {
 /**
  * A minimal / unobtrusive LoadingIndicator displaying an optional spinner and/or message to signal
  * that a longer-running operation is in progress, without using a modal Mask. Can be explicitly
- * shown or bound to one or more tasks.
+ * shown or bound to one or more TaskObservers
  *
  * Note that the Panel component's `loadingIndicator` prop provides a common and convenient way to
  * add an indicator to a Panel without needing to manually create or manage this component.
@@ -42,7 +49,7 @@ export const [LoadingIndicator, loadingIndicator] = hoistCmp.withFactory<Loading
     className: 'xh-loading-indicator',
 
     render(
-        {isDisplayed, message, spinner = true, corner = 'br', maxMessageLength = 30, className},
+        {isDisplayed, message, maxMessageLength = 30, spinner = true, corner = 'br', className},
         ref
     ) {
         const impl = useLocalModel(LocalMaskModel);
@@ -53,19 +60,19 @@ export const [LoadingIndicator, loadingIndicator] = hoistCmp.withFactory<Loading
 
         if (!isDisplayed || (!spinner && !message)) return null;
 
-        const innerItems = () => {
-            const spinnerEl = spinnerCmp({compact: true});
-            if (!message) return [spinnerEl];
-            const msgBox = div({className: `xh-loading-indicator__message`, item: message});
-
-            return corner === 'tl' || corner === 'bl'
-                ? [spinner ? spinnerEl : null, msgBox]
-                : [msgBox, spinner ? spinnerEl : null];
-        };
-
         const hasMessageCls = message ? 'xh-loading-indicator--has-message' : null,
             hasSpinnerCls = spinner ? 'xh-loading-indicator--has-spinner' : null,
             cornerCls = `xh-loading-indicator--${corner}`;
+
+        const innerItems = () => {
+            let spinnerEl = spinnerCmp({compact: true});
+
+            if (!message) return [spinnerEl];
+
+            const msgBox = div({className: `xh-loading-indicator__message`, item: message});
+            if (!spinner) spinnerEl = null;
+            return corner === 'tl' || corner === 'bl' ? [spinnerEl, msgBox] : [msgBox, spinnerEl];
+        };
 
         return div({
             ref,